requirements.txt
requests==2.31.0
flask==2.3.2
WTForms==3.0.1
Django==4.2.4
psycopg2-binary==2.9.7
redis==4.6.0
aiohttp==3.8.5
fastapi==0.101.0
uvicorn==0.23.2
elasticsearch==8.9.0
aioredis==2.0.1
orjson==3.9.4
pytest==7.4.0
pytest-asyncio==0.21.1
jsonschema==4.19.0
grpcio==1.57.0
grpcio-tools==1.57.0
flask_marshmallow==0.15.0
flasgger==0.9.7.1
flask_security==3.0.0
gevent==23.7.0
sqlalchemy==2.0.19
flask_sqlalchemy==3.0.5
pycryptodomex==3.18.0
pyotp==2.9.0
python-multipart==0.0.6
django-model-utils==4.3.1
sentry-sdk[flask]==1.29.2
python-logstash==0.4.8
jinja2==3.1.2
websockets==11.0.3
pika==1.3.2
mypy==1.5.0
